1. Edit the Makeflow user manual 2. Add makeflow_sandbox_delete() in makeflow_gc.c, which is used for deleting makeflow sandbox 3. Merge makeflow_log_clean() into makeflow_log_recover()
