+++ doc/makeflow.html
@@ -42,6 +42,7 @@ based on the image.</p>
 <p>Makeflow currently supports the following batch systems:
 <table>
 <tr><td>local    <td> Local execution on the submitting machine, using multiple cores if available.
+<tr><td>sandbox  <td> Running each task with a sandbox. 
 <tr><td>wq       <td> <a href=http://ccl.cse.nd.edu/software/workqueue>Work Queue</a> distributed application system, included with Makeflow.
 <tr><td>condor   <td> <a href=http://www.cs.wisc.edu/htcondor>HTCondor</a> distributed computing system.
 <tr><td>sge      <td> <a href=http://gridscheduler.sourceforge.net>Open Grid Scheduler</a>, Univa Grid Engine, Oracle Grid Engine, all derived from Sun Grid Engine (SGE).
