+++ doc/man/makeflow.m4
@@ -60,7 +60,7 @@ OPTION_ITEM(`-R, --retry')Automatically retry failed batch jobs up to 100 times.
 OPTION_TRIPLET(-r, retry-count, n)Automatically retry failed batch jobs up to n times.
 OPTION_PAIR(--wait-for-files-upto, #)Wait for output files to be created upto this many seconds (e.g., to deal with NFS semantics).
 OPTION_TRIPLET(-S, submission-timeout, timeout)Time to retry failed batch job submission. (default is 3600s)
-OPTION_TRIPLET(-T, batch-type, type)Batch system type: local, condor, sge, pbs, torque, slurm, moab, cluster, wq, hadoop, mpi-queue. (default is local)
+OPTION_TRIPLET(-T, batch-type, type)Batch system type: local, sandbox, condor, sge, pbs, torque, slurm, moab, cluster, wq, hadoop, mpi-queue. (default is local)
 OPTIONS_END
 
 SUBSECTION(Debugging Options)
