+++ makeflow/src/makeflow_log.h
@@ -27,4 +27,5 @@ void makeflow_log_gc_event( struct dag *d, int collected, timestamp_t elapsed, i
 void makeflow_log_recover( struct dag *d, const char *filename, int verbose_mode, struct batch_queue *queue );
 void makeflow_log_clean( const char *makeflow_log_file );
 void makeflow_log_sandbox_mode_create( struct dag*d, const char *local_task_dir );
+void makeflow_log_sandbox_mode_delete( struct dag*d, const char *local_task_dir );
 #endif
